Move types to higher level directory
