+++ test/types/collection/index.js
