+++ test/types/flag/index.js
