+++ test/types/indexType/index.js
